ingestion_service/requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
pymongo==4.3.3
PyJWT==2.6.0
Flask-Cors==3.0.10
requests==2.32.5